+++ src/Synapse/Synapse.Test/ScenarioTests/SparkPoolTests.ps1
@@ -172,7 +172,7 @@ function Get-WorkspaceTestEnvironmentParameters ($testSuffix)
 			  fileSystemName = "wscmdletfs" + $testSuffix;
 			  loginName = "testlogin";
 			  pwd = "testp@ssMakingIt1007Longer";
-              location = "westcentralus";
+              location = "canadacentral";
               sparkPoolName = "spool" + $testSuffix;
 		}
 }