+++ src/Synapse/Synapse/Commands/ManagementCommands/Workspace/NewAzureSynapseWorkspace.cs
@@ -9,7 +9,7 @@ using Microsoft.Azure.Commands.Synapse.Common;
 using Microsoft.Azure.Commands.Common.Exceptions;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
 
-namespace Microsoft.Azure.Commands.Synaspe
+namespace Microsoft.Azure.Commands.Synapse
 {
     [Cmdlet(VerbsCommon.New, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + SynapseConstants.SynapsePrefix + SynapseConstants.Workspace, SupportsShouldProcess = true)]
     [OutputType(typeof(PSSynapseWorkspace))]
@@ -69,6 +69,12 @@ namespace Microsoft.Azure.Commands.Synaspe
         [Parameter(Mandatory = false, HelpMessage = HelpMessages.AsJob)]
         public SwitchParameter AsJob { get; set; }
 
+        [Parameter(ValueFromPipelineByPropertyName = true, Mandatory = false,
+            HelpMessage = HelpMessages.ManagedResourceGroupName)]
+        [ResourceGroupCompleter()]
+        [ValidateNotNullOrEmpty]
+        public string ManagedResourceGroupName { get; set; }
+
         public override void ExecuteCmdlet()
         {
             try
@@ -80,7 +86,7 @@ namespace Microsoft.Azure.Commands.Synaspe
             }
             catch (AzPSResourceNotFoundCloudException ex)
             {
-                var innerException = ex.InnerException as ErrorContractException;
+                var innerException = ex.InnerException as ErrorResponseException;
                 if (innerException.Body?.Error?.Code == "ResourceNotFound" || innerException.Body?.Error?.Message.Contains("ResourceNotFound") == true)
                 {
                     // account does not exists so go ahead and create one
@@ -117,6 +123,7 @@ namespace Microsoft.Azure.Commands.Synaspe
                 ManagedVirtualNetwork = this.IsParameterBound(c => c.ManagedVirtualNetwork) ? SynapseConstants.DefaultName : null,
                 Location = this.Location,
                 ManagedVirtualNetworkSettings = this.IsParameterBound(c => c.ManagedVirtualNetwork) ? this.ManagedVirtualNetwork?.ToSdkObject() : null,
+                ManagedResourceGroupName = this.ManagedResourceGroupName,
                 Encryption = this.IsParameterBound(c => c.EncryptionKeyIdentifier) ? new EncryptionDetails
                 {
                     Cmk = new CustomerManagedKeyDetails