+++ src/Synapse/Synapse/Models/SynapseConstants.cs
@@ -165,6 +165,12 @@ namespace Microsoft.Azure.Commands.Synapse.Models
 
         public const string SparkJobDefinition = nameof(SparkJobDefinition);
 
+        public const string DisableMaxServiceObjectiveName = "remove";
+
+        public const string ManagedPrivateEndpoint = nameof(ManagedPrivateEndpoint);
+
+        public const string DefaultVNetName = "default";
+
         public static Dictionary<string, ComputeNodeSize> ComputeNodeSizes = new Dictionary<string, ComputeNodeSize>
         {
             {