+++ src/CloudService/Changelog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 
+## Version 0.4.0
+* Introduced 2 additional parameter sets for `New-AzCloudService`.
+
 ## Version 0.3.0
 * Fixed the issue that FrontendIPConfiguration is empty in `Get-AzCloudService`.[#14702]