+++ src/DataFactory/DataFactoryV2/Changelog.md
@@ -18,6 +18,11 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 1.13.1
+* Updated ADF .Net SDK version to 4.21.0
+
+## Version 1.13.0
 * Added Customer Managed Key Encryption to DataFactory
 
 ## Version 1.12.1